feat:Add career lounge, badge feature. Refactor profile feature.

+++ src/main/java/com/project/sanhak/domain/user/UserInfo.java
@@ -9,7 +9,7 @@ public class UserInfo {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long UIid;
+    private int UIid;
 
     @OneToOne
     @JoinColumn(referencedColumnName = "UId")
